Add KoocPollon DM07 + cantbeattacked cond
