+++ game/cards/dm07/fire_bird.go
@@ -1,0 +1,20 @@
+package dm07
+
+import (
+	"duel-masters/game/civ"
+	"duel-masters/game/family"
+	"duel-masters/game/fx"
+	"duel-masters/game/match"
+)
+
+func KoocPollon(c *match.Card) {
+
+	c.Name = "Kooc Pollon"
+	c.Power = 1000
+	c.Civ = civ.Fire
+	c.Family = []string{family.FireBird}
+	c.ManaCost = 2
+	c.ManaRequirement = []string{civ.Fire}
+
+	c.Use(fx.Creature, fx.CantBeAttacked)
+}
